Added test file, renamed surgical fixture to share across both MCODE-specific-procedure-extractors' test files
